fitness_api/requirements.txt
Django
djangorestframework
django-filter

/home/mohamedlazaar360/fitness-tracker-api/fitness_api
fitness_api
venv